feat: transfer parse_command to slack integration
