+++ app/tests/commands/test_utils.py
@@ -29,26 +29,6 @@ def test_log_to_sentinel_logs_error(logging_mock, send_event_mock):
     )
 
 
-def test_parse_command_empty_string():
-    assert utils.parse_command("") == []
-
-
-def test_parse_command_no_args():
-    assert utils.parse_command("sre") == ["sre"]
-
-
-def test_parse_command_one_arg():
-    assert utils.parse_command("sre foo") == ["sre", "foo"]
-
-
-def test_parse_command_two_args():
-    assert utils.parse_command("sre foo bar") == ["sre", "foo", "bar"]
-
-
-def test_parse_command_with_quotes():
-    assert utils.parse_command('sre "foo bar baz"') == ["sre", "foo bar baz"]
-
-
 def test_get_user_locale_supported_locale():
     client = MagicMock()
     user_id = MagicMock()
